% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_classifier.R
\name{column_classifier}
\alias{column_classifier}
\title{Classify the type of each column in a data frame}
\usage{
column_classifier(x)
}
\arguments{
\item{x}{A data frame whose columns are to be classified.}
}
\value{
A data.frame where each row contains the column name and its type.
}
\description{
This function takes a data frame as input and returns a data frame
where each row corresponds to a column in the input.
}
