% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_sig}
\alias{format_sig}
\title{Format p-values with significance stars}
\usage{
format_sig(
  x,
  k = 3,
  thresholds = c(0.001, 0.05, 0.1),
  stars = c("***", "**", "*", "")
)
}
\arguments{
\item{x}{Numeric p-value.}

\item{k}{Number of decimal places to round.}

\item{thresholds}{Numeric vector of significance limits.}

\item{stars}{Character vector of symbols corresponding to each threshold.}
}
\value{
Character string of the rounded p-value and stars.
}
\description{
Add significance stars to a numeric value according to provided thresholds.
}
\examples{
format_sig(0.02)
}
