% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd1.R
\name{sd1}
\alias{sd1}
\title{Safe standard deviation}
\usage{
sd1(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric standard deviation or 0.
}
\description{
Compute the standard deviation of a numeric vector returning 0 when length is less than two.
}
\examples{
sd1(c(1, 2, 3))
}
