% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_freq.R
\name{tab_freq}
\alias{tab_freq}
\title{Frequency table}
\usage{
tab_freq(aux)
}
\arguments{
\item{aux}{Data frame containing a \code{.y.} column.}
}
\value{
Tibble with counts and percentages for each level.
}
\description{
Create a simple frequency table for the column \code{.y.} of a data frame.
}
\examples{
tab_freq(tibble::tibble(.y. = c('A','B','A')))
}
