% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{msdr}
\alias{msdr}
\title{Numeric summary string}
\usage{
msdr(x, k = 2)
}
\arguments{
\item{x}{Numeric vector to summarise.}

\item{k}{Number of decimal places to round to.}
}
\value{
A character string in the form "mean plus and minus sd (min~max)".
}
\description{
Format the mean, standard deviation and range of a numeric vector.
}
\examples{
msdr(c(1, 2, 3))
}
